chore: Moved config to internal
