+++ internal/world/world.go
@@ -4,7 +4,7 @@ import (
 	"math"
 
 	"github.com/N3moAhead/harvest/internal/component"
-	"github.com/N3moAhead/harvest/pkg/config"
+	"github.com/N3moAhead/harvest/internal/config"
 	"github.com/hajimehoshi/ebiten/v2"
 )
