+++ internal/world/worldgenerator.go
@@ -5,7 +5,7 @@ import (
 	"math/rand"
 
 	"github.com/N3moAhead/harvest/internal/assets"
-	"github.com/N3moAhead/harvest/pkg/config"
+	"github.com/N3moAhead/harvest/internal/config"
 	"github.com/hajimehoshi/ebiten/v2"
 )
